+++ src/plonemeeting/portal/core/config.py
@@ -33,7 +33,7 @@ CATEGORY_IA_DELIB_FIELDS_MAPPING_EXTRA_INCLUDE = {
     "classifier": "classifiers",
 }
 
-REPRESENTATIVE_IA_DELIB_FIELD = 'groups_in_charge'
+REPRESENTATIVE_IA_DELIB_FIELD = "groups_in_charge"
 
 
 REGION_INS_CODE = "03000"
@@ -44,6 +44,24 @@ DEMO_INSTITUTION_IDS = ["belle-ville"]
 FACETED_DEC_MANAGER_CRITERIA = ["annexes"]
 
 FACETED_PUB_MANAGER_CRITERIA = ["annexes", "etat"]
+
+DOCUMENTGENERATOR_GENERABLE_CONTENT_TYPES = ["Meeting", "Institution", "Item", "Publication", "Folder"]
+DOCUMENTGENENATOR_USED_CONTENT_TYPES = ["ConfigurablePODTemplate", "PODTemplate", "StyleTemplate", "SubTemplate"]
+DEFAULT_DOCUMENTGENERATOR_TEMPLATES = {
+    "publication_summary": {
+        "title": _("Publication summary"),
+        "odt_file": "publication_summary.odt",
+        "pod_formats": ["pdf"],
+        "pod_portal_types": ["Publication"],
+    },
+    "publication_public_notice": {
+        "title": _("Publication public notice"),
+        "odt_file": "publication_public_notice.odt",
+        "pod_formats": ["pdf"],
+        "pod_portal_types": ["Publication"],
+    },
+}
+
 
 MIMETYPE_TO_ICON = {
     "text/plain": {"icon": "bi bi-file-text", "color": "bg-light"},