+++ src/plonemeeting/portal/core/patches.py
@@ -12,6 +12,7 @@ from plone.base.navigationroot import get_navigation_root_object
 from plone.base.utils import get_top_site_from_url
 from plone.supermodel.interfaces import FIELDSETS_KEY
 from plonemeeting.portal.core import logger
+from plonemeeting.portal.core.interfaces import IPlonemeetingPortalCoreLayer
 from z3c.form.interfaces import IForm
 from zope.globalrequest import getRequest
 
@@ -59,6 +60,9 @@ def get_contentbrowser_options(*args, **kwargs):
     context = args[0] if args else kwargs.get("context")
     if IForm.providedBy(context):  # Sometimes context is a form
         context = context.context
+    if not IPlonemeetingPortalCoreLayer.implementedBy(context):
+        # Don't break if plonemeeting.portal.core is not installed
+        return res
     request = getRequest()
     site = get_top_site_from_url(context, request)
     nav_root = get_navigation_root_object(context, site)